fix qwtplot3d for Qt5

+++ qwtplot3d/src/qwt3d_lighting.cpp
@@ -175,7 +175,8 @@ void Plot3D::applyLight(unsigned light)
   glRotatef( lights_[light].rot.x-90, 1.0, 0.0, 0.0 ); 
   glRotatef( lights_[light].rot.y   , 0.0, 1.0, 0.0 ); 
   glRotatef( lights_[light].rot.z   , 0.0, 0.0, 1.0 );
-  GLfloat lightPos[4] = { lights_[light].shift.x, lights_[light].shift.y, lights_[light].shift.z, 1.0};
+  //GLfloat lightPos[4] = { lights_[light].shift.x, lights_[light].shift.y, lights_[light].shift.z, 1.0};
+  GLfloat lightPos[4] = { (GLfloat)lights_[light].shift.x, (GLfloat)lights_[light].shift.y, (GLfloat)lights_[light].shift.z, (GLfloat)1.0};
   GLenum le = lightEnum(light);
   glLightfv(le, GL_POSITION, lightPos);  
 }
